bash customizer.sh com.exapmle.ogasample DataItemType OgaSample
