+++ app/src/androidTest/java/com/exapmle/ogasample/testdi/TestDatabaseModule.kt
@@ -14,20 +14,25 @@
  * limitations under the License.
  */
 
-pluginManagement {
-    repositories {
-        gradlePluginPortal()
-        google()
-        mavenCentral()
-    }
-}
-dependencyResolutionManagement {
-    repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
-    repositories {
-        google()
-        mavenCentral()
-    }
-}
-rootProject.name = "OgaSample"
+package com.exapmle.ogasample.testdi
 
-include(":app")
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.components.SingletonComponent
+import dagger.hilt.testing.TestInstallIn
+import com.exapmle.ogasample.data.DataItemTypeRepository
+import com.exapmle.ogasample.data.di.DataModule
+import com.exapmle.ogasample.data.di.FakeDataItemTypeRepository
+
+@Module
+@TestInstallIn(
+    components = [SingletonComponent::class],
+    replaces = [DataModule::class]
+)
+interface FakeDataModule {
+
+    @Binds
+    abstract fun bindRepository(
+        fakeRepository: FakeDataItemTypeRepository
+    ): DataItemTypeRepository
+}
