+++ app/src/androidTest/java/com/exapmle/ogasample/ui/dataitemtype/DataItemTypeScreenTest.kt
@@ -1,0 +1,51 @@
+/*
+ * Copyright (C) 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.exapmle.ogasample.ui.dataitemtype
+
+import androidx.activity.ComponentActivity
+import androidx.compose.ui.test.junit4.createAndroidComposeRule
+import androidx.compose.ui.test.onNodeWithText
+import androidx.compose.ui.test.performClick
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import org.junit.Before
+import org.junit.Rule
+import org.junit.Test
+import org.junit.runner.RunWith
+
+/**
+ * UI tests for [DataItemTypeScreen].
+ */
+@RunWith(AndroidJUnit4::class)
+class DataItemTypeScreenTest {
+
+    @get:Rule
+    val composeTestRule = createAndroidComposeRule<ComponentActivity>()
+
+    @Before
+    fun setup() {
+        composeTestRule.setContent {
+            DataItemTypeScreen(FAKE_DATA, onSave = {})
+        }
+    }
+
+    @Test
+    fun firstItem_exists() {
+        composeTestRule.onNodeWithText(FAKE_DATA.first()).assertExists().performClick()
+    }
+}
+
+private val FAKE_DATA = listOf("Compose", "Room", "Kotlin")
