--- app/src/main/java/android/template/data/MyModelRepository.kt
@@ -1,41 +1,0 @@
-/*
- * Copyright (C) 2022 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package android.template.data
-
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
-import android.template.data.local.database.MyModel
-import android.template.data.local.database.MyModelDao
-import javax.inject.Inject
-
-interface MyModelRepository {
-    val myModels: Flow<List<String>>
-
-    suspend fun add(name: String)
-}
-
-class DefaultMyModelRepository @Inject constructor(
-    private val myModelDao: MyModelDao
-) : MyModelRepository {
-
-    override val myModels: Flow<List<String>> =
-        myModelDao.getMyModels().map { items -> items.map { it.name } }
-
-    override suspend fun add(name: String) {
-        myModelDao.insertMyModel(MyModel(name = name))
-    }
-}
