+++ app/src/main/java/com/exapmle/ogasample/ui/dataitemtype/DataItemTypeScreen.kt
@@ -1,0 +1,109 @@
+/*
+ * Copyright (C) 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.exapmle.ogasample.ui.dataitemtype
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.produceState
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalLifecycleOwner
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.Lifecycle.State.STARTED
+import androidx.lifecycle.repeatOnLifecycle
+import com.exapmle.ogasample.ui.theme.MyApplicationTheme
+import androidx.compose.material3.ExperimentalMaterial3Api
+
+@Composable
+fun DataItemTypeScreen(modifier: Modifier = Modifier, viewModel: DataItemTypeViewModel = hiltViewModel()) {
+    val lifecycle = LocalLifecycleOwner.current.lifecycle
+    val items by produceState<DataItemTypeUiState>(
+        initialValue = DataItemTypeUiState.Loading,
+        key1 = lifecycle,
+        key2 = viewModel
+    ) {
+        lifecycle.repeatOnLifecycle(state = STARTED) {
+            viewModel.uiState.collect { value = it }
+        }
+    }
+    if (items is DataItemTypeUiState.Success) {
+        DataItemTypeScreen(
+            items = (items as DataItemTypeUiState.Success).data,
+            onSave = viewModel::addDataItemType,
+            modifier = modifier
+        )
+    }
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+internal fun DataItemTypeScreen(
+    items: List<String>,
+    onSave: (name: String) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Column(modifier) {
+        var nameDataItemType by remember { mutableStateOf("Compose") }
+        Row(
+            modifier = Modifier.fillMaxWidth().padding(bottom = 24.dp),
+            horizontalArrangement = Arrangement.spacedBy(16.dp)
+        ) {
+            TextField(
+                value = nameDataItemType,
+                onValueChange = { nameDataItemType = it }
+            )
+
+            Button(modifier = Modifier.width(96.dp), onClick = { onSave(nameDataItemType) }) {
+                Text("Save")
+            }
+        }
+        items.forEach {
+            Text("Saved item: $it")
+        }
+    }
+}
+
+// Previews
+
+@Preview(showBackground = true)
+@Composable
+private fun DefaultPreview() {
+    MyApplicationTheme {
+        DataItemTypeScreen(listOf("Compose", "Room", "Kotlin"), onSave = {})
+    }
+}
+
+@Preview(showBackground = true, widthDp = 480)
+@Composable
+private fun PortraitPreview() {
+    MyApplicationTheme {
+        DataItemTypeScreen(listOf("Compose", "Room", "Kotlin"), onSave = {})
+    }
+}
