+++ app/src/main/res/values/strings.xml
@@ -15,5 +15,5 @@
   -->
 
 <resources>
-    <string name="app_name">MyApplication</string>
+    <string name="app_name">OgaSample</string>
 </resources>
